Rename app to Android Service Delivery Demo
