+++ src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.fidesmo.sec.transceivedemo"
+    package="com.fidesmo.sec.servicedeliverydemo"
     android:versionCode="1"
     android:versionName="1.0" >
 
@@ -14,7 +14,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name="com.fidesmo.sec.transceivedemo.MainActivity"
+            android:name="com.fidesmo.sec.servicedeliverydemo.MainActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait" >
             <intent-filter>
@@ -23,12 +23,12 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.fidesmo.sec.transceivedemo.TransceiveActivity"
+            android:name="com.fidesmo.sec.servicedeliverydemo.TransceiveActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait">
         </activity>
         <activity
-            android:name="com.fidesmo.sec.transceivedemo.ServiceDeliveryActivity"
+            android:name="com.fidesmo.sec.servicedeliverydemo.ServiceDeliveryActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait">
         </activity>
